requirements.txt
# Core Inspect AI dependencies
inspect_ai>=0.3.0

# Data and analysis
pandas>=2.0.0
datasets>=3.0.0
tqdm>=4.60.0
jsonlines>=4.0.0
PyYAML>=6.0.0

# Utilities
platformdirs>=4.0.0

# Optional extras:
# - SWE support: install from setup extras `pip install -e .[swe]`
# - LLM judge support: install from setup extras `pip install -e .[analysis]`
